requirements.txt
amqp==2.6.1
asgiref==3.5.2
billiard==3.6.4.0
celery==4.4.7
Django==3.2.9
django-celery-beat==2.1.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
kombu==4.6.11
Markdown==3.3.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.5
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
vine==1.3.0